--- sharding-transaction/sharding-transaction-xa/src/main/java/io/shardingsphere/transaction/manager/xa/convert/Converter.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2016-2018 shardingsphere.io.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * </p>
- */
-
-package io.shardingsphere.transaction.manager.xa.convert;
-
-import io.shardingsphere.core.rule.DataSourceParameter;
-
-/**
- * Convert interface.
- *
- * @author zhaojun
- */
-public interface Converter {
-    
-    /**
-     * Get the DataSource pool property, then convert to datasource parameter.
-     *
-     * @return dataSource parameter
-     */
-    DataSourceParameter convertTo();
-}
-
