+++ sharding-transaction/sharding-transaction-xa/src/main/java/io/shardingsphere/transaction/xa/manager/property/MysqlXAPropertyFactory.java
@@ -1,0 +1,61 @@
+/*
+ * Copyright 2016-2018 shardingsphere.io.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * </p>
+ */
+
+package io.shardingsphere.transaction.xa.manager.property;
+
+import com.google.common.base.Optional;
+import io.shardingsphere.core.rule.DataSourceParameter;
+import lombok.RequiredArgsConstructor;
+
+import java.util.Properties;
+
+/**
+ * Create Mysql XA property from data source parameter.
+ *
+ * @author zhaojun
+ */
+@RequiredArgsConstructor
+public class MysqlXAPropertyFactory {
+    
+    private final DataSourceParameter dataSourceParameter;
+    
+    /**
+     * Build Mysql XA properties.
+     *
+     * @return Mysql XA properties
+     */
+    public Properties build() {
+        Properties result = new Properties();
+        result.setProperty("user", dataSourceParameter.getUsername());
+        result.setProperty("password", Optional.fromNullable(dataSourceParameter.getPassword()).or(""));
+        result.setProperty("URL", dataSourceParameter.getUrl());
+        result.setProperty("pinGlobalTxToPhysicalConnection", Boolean.TRUE.toString());
+        result.setProperty("autoReconnect", Boolean.TRUE.toString());
+        result.setProperty("useServerPrepStmts", Boolean.TRUE.toString());
+        result.setProperty("cachePrepStmts", Boolean.TRUE.toString());
+        result.setProperty("prepStmtCacheSize", "250");
+        result.setProperty("prepStmtCacheSqlLimit", "2048");
+        result.setProperty("useLocalSessionState", Boolean.TRUE.toString());
+        result.setProperty("rewriteBatchedStatements", Boolean.TRUE.toString());
+        result.setProperty("cacheResultSetMetadata", Boolean.TRUE.toString());
+        result.setProperty("cacheServerConfiguration", Boolean.TRUE.toString());
+        result.setProperty("elideSetAutoCommits", Boolean.TRUE.toString());
+        result.setProperty("maintainTimeStats", Boolean.FALSE.toString());
+        result.setProperty("netTimeoutForStreamingResults", "0");
+        return result;
+    }
+}
