+++ sharding-transaction/sharding-transaction-xa/src/main/java/io/shardingsphere/transaction/xa/manager/property/XAPropertyFactory.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2016-2018 shardingsphere.io.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * </p>
+ */
+
+package io.shardingsphere.transaction.xa.manager.property;
+
+import io.shardingsphere.core.rule.DataSourceParameter;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+import java.util.Properties;
+
+/**
+ * XA property factory.
+ *
+ * @author zhaojun
+ */
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class XAPropertyFactory {
+    
+    /**
+     * Create XA properties using datasource parameter.
+     *
+     * @param xaDatabaseType XA database type
+     * @param dataSourceParameter datasource parameter
+     * @return XA properties
+     */
+    public static Properties build(final XADatabaseType xaDatabaseType, final DataSourceParameter dataSourceParameter) {
+        switch (xaDatabaseType) {
+            case MySQL:
+                return new MysqlXAPropertyFactory(dataSourceParameter).build();
+            default:
+                return new Properties();
+        }
+    }
+}
