+++ sharding-transaction/sharding-transaction-xa/src/main/resources/META-INF/services/io.shardingsphere.spi.transaction.ShardingTransactionHandler
@@ -1,2 +1,2 @@
-io.shardingsphere.transaction.handler.xa.XAShardingTransactionHandler
+io.shardingsphere.transaction.xa.handler.XAShardingTransactionHandler
