+++ sharding-transaction/sharding-transaction-xa/src/test/resources/META-INF/services/io.shardingsphere.spi.xa.BackendDataSourceFactory
@@ -1,1 +1,1 @@
-io.shardingsphere.transaction.manager.xa.XABackendDataSourceConverter
+io.shardingsphere.transaction.xa.manager.XABackendDataSourceConverter
