--- sharding-transaction/sharding-transaction-xa/src/test/resources/META-INF/services/io.shardingsphere.transaction.manager.xa.XATransactionManager
@@ -1,2 +1,0 @@
-io.shardingsphere.transaction.manager.xa.atomikos.AtomikosTransactionManager
-
