+++ sharding-transaction/sharding-transaction-xa/src/test/resources/META-INF/services/io.shardingsphere.transaction.xa.manager.XATransactionManager
@@ -1,0 +1,2 @@
+io.shardingsphere.transaction.xa.manager.atomikos.AtomikosTransactionManager
+
